Add files via upload
The mimeTypes.rdf file is  responsible for how Mozilla Firefox handles the downloads of all file types on the internet. The file is located in the Mozilla Firefox profile folder which is usually located in the following directory:

C:\Users\Computer_User_Name\AppData\Roaming\Mozilla\Firefox\Profiles\Folder_Name.default

This file is to be downloaded and placed in the above folder overwriting the existing mime type.rdf file.